# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/post": {
      "get": {
        "operationId": "PostController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of posts to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of posts to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagIds",
            "required": false,
            "in": "query",
            "description": "Filter by tag IDs (comma-separated)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in title and content",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of posts with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all posts with optional filters",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new post",
        "tags": [
          "posts"
        ]
      }
    },
    "/post/{id}": {
      "get": {
        "operationId": "PostController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return post details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get post by ID",
        "tags": [
          "posts"
        ]
      },
      "put": {
        "operationId": "PostController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to update post"
          },
          "404": {
            "description": "Post not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a post",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to delete post"
          },
          "404": {
            "description": "Post not found"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/post/{id}/lock": {
      "put": {
        "operationId": "PostController_toggleLock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post lock status toggled successfully"
          },
          "403": {
            "description": "Not authorized to lock/unlock post"
          },
          "404": {
            "description": "Post not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Toggle post lock status (Admin only)",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/notifications": {
      "post": {
        "operationId": "NotificationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thông báo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo thông báo mới",
        "tags": [
          "notifications"
        ]
      },
      "get": {
        "operationId": "NotificationController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách thông báo"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách thông báo của người dùng hiện tại",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/unread/count": {
      "get": {
        "operationId": "NotificationController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Số lượng thông báo chưa đọc"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đếm số thông báo chưa đọc",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đánh dấu đã đọc thành công"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đánh dấu thông báo đã đọc",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/read/all": {
      "patch": {
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Đánh dấu tất cả đã đọc thành công"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đánh dấu tất cả thông báo đã đọc",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "delete": {
        "operationId": "NotificationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thông báo thành công"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa thông báo",
        "tags": [
          "notifications"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "The API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string",
            "description": "User role"
          }
        },
        "required": [
          "_id",
          "email",
          "username",
          "password",
          "fullname",
          "role"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Vote": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của lượt bình chọn"
          },
          "user": {
            "description": "Người dùng thực hiện bình chọn",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của người dùng thực hiện bình chọn"
          },
          "targetType": {
            "enum": [
              "post",
              "comment"
            ],
            "type": "string",
            "description": "Loại đối tượng được bình chọn (bài viết hoặc bình luận)",
            "example": "post"
          },
          "targetId": {
            "type": "string",
            "description": "ID của đối tượng được bình chọn"
          },
          "voteValue": {
            "enum": [
              1,
              -1
            ],
            "type": "number",
            "description": "Giá trị bình chọn (1 cho upvote, -1 cho downvote)",
            "example": 1
          },
          "post": {
            "description": "Bài viết được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "comment": {
            "description": "Bình luận được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình chọn"
          }
        },
        "required": [
          "_id",
          "user",
          "userId",
          "targetType",
          "targetId",
          "voteValue",
          "createdAt"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bình luận"
          },
          "post": {
            "description": "Bài viết chứa bình luận này",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "postId": {
            "type": "string",
            "description": "ID của bài viết"
          },
          "user": {
            "description": "Tác giả bình luận",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          },
          "parentComment": {
            "description": "Bình luận cha (nếu là bình luận trả lời)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "parentCommentId": {
            "type": "string",
            "description": "ID của bình luận cha (nếu có)"
          },
          "replies": {
            "description": "Các bình luận trả lời",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bình luận",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình luận"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bình luận"
          }
        },
        "required": [
          "_id",
          "post",
          "postId",
          "user",
          "userId",
          "content",
          "replies",
          "votes",
          "createdAt",
          "updatedAt"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bài viết"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "user": {
            "description": "Tác giả bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "category": {
            "description": "Danh mục của bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID của danh mục"
          },
          "tags": {
            "description": "Các thẻ được gắn cho bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "comments": {
            "description": "Các bình luận của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "isLocked": {
            "type": "boolean",
            "description": "Trạng thái khóa của bài viết"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bài viết"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bài viết"
          }
        },
        "required": [
          "_id",
          "title",
          "content",
          "user",
          "userId",
          "category",
          "categoryId",
          "tags",
          "comments",
          "votes",
          "isLocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của thẻ"
          },
          "tagName": {
            "type": "string",
            "description": "Tên của thẻ",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "Mô tả của thẻ"
          },
          "posts": {
            "description": "Các bài viết được gắn thẻ này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          }
        },
        "required": [
          "_id",
          "tagName",
          "posts"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "categoryId": {
            "type": "string",
            "description": "ID của danh mục"
          },
          "tagIds": {
            "description": "Danh sách ID của các thẻ",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "content",
          "categoryId"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "tagIds": {
            "description": "Danh sách ID của các thẻ",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isLocked": {
            "type": "boolean",
            "description": "Trạng thái khóa của bài viết"
          }
        }
      },
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID người dùng nhận thông báo"
          },
          "type": {
            "enum": [
              "post_created",
              "comment_created",
              "post_voted",
              "comment_voted",
              "post_reported",
              "comment_reported",
              "report_resolved"
            ],
            "type": "string",
            "description": "Loại thông báo"
          },
          "content": {
            "type": "string",
            "description": "Nội dung thông báo"
          }
        },
        "required": [
          "userId",
          "type",
          "content"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID của thông báo"
          },
          "user": {
            "description": "Người dùng nhận thông báo",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID người dùng nhận thông báo"
          },
          "type": {
            "enum": [
              "post_created",
              "comment_created",
              "post_voted",
              "comment_voted",
              "post_reported",
              "comment_reported",
              "report_resolved"
            ],
            "type": "string",
            "description": "Loại thông báo"
          },
          "content": {
            "type": "string",
            "description": "Nội dung thông báo"
          },
          "isRead": {
            "type": "boolean",
            "description": "Trạng thái đã đọc"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời gian tạo"
          }
        },
        "required": [
          "_id",
          "user",
          "userId",
          "type",
          "content",
          "isRead",
          "createdAt"
        ]
      }
    }
  }
}